refactor(TicketStatusLabel): extract to own file
