+++ modules/Tickets.js
@@ -8,9 +8,10 @@ import AV from 'leancloud-storage/live-query'
 import css from './CustomerServiceTickets.css'
 import DocumentTitle from 'react-document-title'
 
-import {UserLabel, TicketStatusLabel, getCategoryPathName, getCategoriesTree, getTicketAcl} from './common'
+import {UserLabel, getCategoryPathName, getCategoriesTree, getTicketAcl} from './common'
 import OrganizationSelect from './OrganizationSelect'
 import TicketsMoveButton from './TicketsMoveButton'
+import TicketStatusLabel from './TicketStatusLabel'
 import translate from './i18n/translate'
 
 class Tickets extends Component {
